--- monitoring-core/src/main/java/com/monikit/core/HttpOutboundRequestLog.java
@@ -1,50 +1,0 @@
-package com.monikit.core;
-
-import java.util.Map;
-
-/**
- * 내 서버에서 외부 API로 요청을 보낼 때 실행된 HTTP 요청을 기록하는 로그 클래스.
- * <p>
- * 외부 API URL, HTTP 메서드, 요청 헤더, 요청 바디 등을 포함하여
- * API 연동 요청 이력을 추적할 수 있다.
- * </p>
- *
- * @author ryu-qqq
- * @since 1.0
- */
-public class HttpOutboundRequestLog extends AbstractLogEntry {
-    private final String httpMethod;
-    private final String targetUrl;
-    private final String headers;
-    private final String requestBody;
-    private final long executionTime;
-
-    protected HttpOutboundRequestLog(String traceId, String httpMethod, String targetUrl, String headers,
-                                     String requestBody, long executionTime, LogLevel logLevel) {
-        super(traceId, logLevel);
-        this.httpMethod = httpMethod;
-        this.targetUrl = targetUrl;
-        this.headers = headers;
-        this.requestBody = requestBody;
-        this.executionTime = executionTime;
-    }
-
-    @Override
-    public LogType getLogType() {
-        return LogType.OUTBOUND_REQUEST;
-    }
-
-    @Override
-    protected void addExtraFields(Map<String, Object> logMap) {
-        logMap.put("httpMethod", httpMethod);
-        logMap.put("targetUrl", targetUrl);
-        logMap.put("headers", headers);
-        logMap.put("requestBody", requestBody);
-        logMap.put("executionTime", executionTime + "ms");
-    }
-
-    public static HttpOutboundRequestLog create(String traceId, String httpMethod, String targetUrl, String headers,
-                                                String requestBody, long executionTime, LogLevel logLevel) {
-        return new HttpOutboundRequestLog(traceId, httpMethod, targetUrl, headers, requestBody, executionTime, logLevel);
-    }
-}
